Refactor: restructure folders into dao/domain/manualTest
