+++ src/main/java/app/TransactionHistoryApp.java
@@ -1,11 +1,11 @@
 package app;
 
-import account.dao.AccountDaoImpl;
+import dao.AccountDaoImpl;
 import service.transaction.TransactionService;
 import service.transaction.TransactionServiceImpl;
-import transaction.dao.TransactionDao;
-import transaction.dao.TransactionDaoImpl;
-import transaction.domain.TransactionVO;
+import dao.TransactionDao;
+import dao.TransactionDaoImpl;
+import domain.TransactionVO;
 
 import java.math.BigDecimal;
 import java.time.format.DateTimeFormatter;
