+++ src/main/java/domain/ProductVO.java
@@ -1,4 +1,4 @@
-package product.domain;
+package domain;
 
 
 import lombok.AllArgsConstructor;
