+++ src/main/java/manualTest/user_productDaoManualTest.java
@@ -1,8 +1,8 @@
-package user_product;
+package manualTest;
 
-import user_product.dao.user_productDao;
-import user_product.dao.user_productDaoImpl;
-import user_product.domain.user_productVO;
+import dao.user_productDao;
+import dao.user_productDaoImpl;
+import domain.user_productVO;
 
 import java.time.LocalDate;
 import java.util.List;
