+++ src/main/java/service/productList/ProductListService.java
@@ -1,6 +1,6 @@
 package service.productList;
 
-import user_product.domain.user_productVO;
+import domain.user_productVO;
 
 import java.util.List;
